Specify port when calling all smbclient methods
If a non-standard (_ie_ not 445) port is configured for the SMB server,
we must specify the port explicitly when calling various smbclient
methods. Otherwise, smbclient wrongly assumes port 445, even though we
have earlier specified the port when calling `register_session`.

It's possible that with this change we are working around a bug in
smbclient, and one that isn't frequently encountered, since it seems
likely that people would usually stick to the standard port.

Either way, this change should be safe, and should ensure that the
network drive connector will work in the rare cases where port 445 is
not being used.